third_party/blink/web_tests/fast/dom/custom/callback-timing-expected.txt

CONSOLE WARNING: line 22: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Ensure that custom element callback fires before DOMContentLoaded event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS eventlog[0] is 'created x-foo'
PASS eventlog[1] is 'DOMContentLoaded event'
PASS eventlog[2] is 'load event'
